Compile and Simulation added. Compile is handled and job files get created. Compile updates the database on file situation. Compiled files display their situation to the directory list. Compile errors are shown from log file if the action has failed. Simulation form constructed. Simulation hadnled and job files get created. Other minor fixes.

+++ pages/extra/list-dirfiles.php
@@ -10,7 +10,25 @@
 			$path = $BASE_URL."/project/".$search_user['username']."/".$_GET['project']."/".$file['type'].$file['relative_path']."/".$file['name'];
 		}
 		echo "<li class='list-group-item'><a href='".$path."'>".$file['name']."</a>";
-		echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+		if($file['compiled']=='1'){
+			$full_path = $BASE_DIR.$search_user['username'].'/'.$_GET['project'].'/'.$file['name'];
+			$log_file = $full_path.".log";
+			if( file_exists($log_file) ){
+				if( filesize($log_file) == 0 ){
+					echo "<span class='compile-info compile-success'>Compiled</span>";
+					echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+				}else{
+					echo "<span class='compile-info compile-error'><a href='#error_compile_".$file['id']."' data-toggle='collapse'> Compile Failed <span class='glyphicon glyphicon-chevron-down'></span></a></span>";
+					echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+					echo "<div id='error_compile_".$file['id']."' class='collapse'><pre class='alert-danger'>".file_get_contents($log_file)."</pre></div>";
+				}
+			}else{
+				echo "<span class='compile-info compile-pending'> Pending Compile</span>";
+				echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+			}
+		}else{
+			echo "<span class='pull-right'><input type='checkbox' class='select_file' value='".$file['id']."' id='file_".$file['id']."'></span></li>";
+		}
 	}
 	if($user->validate_edit_rights($editors)){ ?>
 		<li class='list-group-item'>
